requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==3.2.13
django-rest-knox==4.2.0
djangorestframework==3.13.1
Pillow==9.1.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
